+++ packages/desktop/src/renderer/components/widgets/channels/ChannelInput/ChannelInput.tsx
@@ -1,6 +1,5 @@
 import React, { ReactElement, useCallback } from 'react'
 import classNames from 'classnames'
-import ContentEditable, { ContentEditableEvent } from 'react-contenteditable'
 import Picker, { EmojiStyle } from 'emoji-picker-react'
 import Grid from '@mui/material/Grid'
 import { styled } from '@mui/material/styles'
@@ -14,7 +13,6 @@ import emojiBlack from '../../../../static/images/emojiBlack.svg'
 import paperclipGray from '../../../../static/images/paperclipGray.svg'
 import paperclipBlack from '../../../../static/images/paperclipBlack.svg'
 import path from 'path'
-import { isDefined } from '@quiet/common'
 
 const PREFIX = 'ChannelInput'
 
@@ -41,7 +39,6 @@ const classes = {
   notAllowed: `${PREFIX}notAllowed`,
   inputFiles: `${PREFIX}inputFiles`,
   icons: `${PREFIX}icons`,
-  textArea: `${PREFIX}textArea`,
 }
 
 const maxHeight = 300
@@ -62,11 +59,17 @@ const StyledChannelInput = styled(Grid)(({ theme }) => ({
     to: { opacity: 1 },
   },
   [`& .${classes.input}`]: {
+    display: 'block',
+    border: 0,
+    resize: 'none',
+    fontFamily: 'inherit',
     whiteSpace: 'break-spaces',
     width: '100%',
     fontSize: 14,
     outline: 'none',
     padding: '12px 16px',
+    scrollPaddingBottom: '12px',
+    height: '48px',
     lineHeight: '24px',
     '&:empty': {
       '&:before': {
@@ -74,6 +77,9 @@ const StyledChannelInput = styled(Grid)(({ theme }) => ({
         display: 'block',
         color: '#aaa',
       },
+    },
+    '&::placeholder': {
+      color: '#aaa',
     },
     wordBreak: 'break-word',
     position: 'relative',
@@ -200,15 +206,11 @@ const StyledChannelInput = styled(Grid)(({ theme }) => ({
     alignItems: 'center',
     alignCntent: 'stretch',
   },
-  [`& .${classes.textArea}`]: {
-    maxHeight: maxHeight,
-  },
 }))
 
 export interface ChannelInputProps {
   channelId: string
   channelName?: string
-  channelParticipants?: Array<{ nickname: string }>
   inputPlaceholder: string
   inputState?: INPUT_STATE
   initialMessage?: string
@@ -224,7 +226,6 @@ export interface ChannelInputProps {
 
 export const ChannelInputComponent: React.FC<ChannelInputProps> = ({
   channelId,
-  channelParticipants = [],
   inputPlaceholder,
   inputState = INPUT_STATE.AVAILABLE,
   initialMessage = '',
@@ -237,114 +238,36 @@ export const ChannelInputComponent: React.FC<ChannelInputProps> = ({
   handleClipboardFiles,
   handleOpenFiles,
 }) => {
-  const [_anchorEl, setAnchorEl] = React.useState<HTMLDivElement>()
-  const [mentionsToSelect, setMentionsToSelect] = React.useState<
-    Array<{
-      nickname: string
-    }>
-  >([])
-  const messageRef = React.useRef<string>()
-  const refSelected = React.useRef<number>()
-  const isFirstRenderRef = React.useRef(true)
-
-  const mentionsToSelectRef = React.useRef<any[]>()
-
-  const inputRef = React.createRef<ContentEditable & HTMLDivElement & any>() // any for updater.enqueueForceUpdate
-
+  const textAreaRef = React.createRef<HTMLTextAreaElement>()
   const fileInput = React.useRef<HTMLInputElement>(null)
 
   const [focused, setFocused] = React.useState(false)
-  const [selected, setSelected] = React.useState(0)
 
   const [emojiHovered, setEmojiHovered] = React.useState(false)
   const [fileExplorerHovered, setFileExplorerHovered] = React.useState(false)
   const [openEmoji, setOpenEmoji] = React.useState(false)
 
-  const [htmlMessage, setHtmlMessage] = React.useState<string>(initialMessage)
   const [message, setMessage] = React.useState(initialMessage)
 
   React.useEffect(() => {
-    inputRef.current?.el.current.focus()
-  }, [inputRef])
-
-  React.useEffect(() => {
-    inputRef.current.updater.enqueueForceUpdate(inputRef.current)
-  }, [inputPlaceholder, channelId])
-
-  // Use reference to bypass memorization
-  React.useEffect(() => {
-    refSelected.current = selected
-  }, [selected])
-
-  const isRefSelected = (refSelected: number | undefined): refSelected is number => {
-    return typeof refSelected === 'number'
-  }
-
-  React.useEffect(() => {
-    mentionsToSelectRef.current = mentionsToSelect
-  }, [mentionsToSelect])
-
-  React.useEffect(() => {
-    if (!message) {
-      setHtmlMessage('')
-    }
-  }, [message])
-
-  React.useEffect(() => {
     setMessage(initialMessage)
-    setHtmlMessage(initialMessage)
-    if (!isFirstRenderRef.current) {
-      return () => {
-        if (messageRef?.current) {
-          onChange(messageRef.current)
-        }
-      }
-    }
-    isFirstRenderRef.current = false
+    const ref = textAreaRef.current
+    if (!ref) return
+    if (!initialMessage) return
+    adjustTextAreaHeight(ref)
   }, [channelId])
 
   React.useEffect(() => {
-    messageRef.current = message
-  }, [message])
+    textAreaRef.current?.focus()
+  }, [textAreaRef])
 
-  const findMentions = React.useCallback(
-    (text: string) => {
-      // Search for any mention in message string
-      const result: string = text.replace(
-        /(<span([^>]*)>)?@([a-z0-9]?\w*)(<\/span>)?/gi,
-        (match, span, _class, nickname) => {
-          // Ignore already established mentions
-          if (span?.includes('class')) {
-            return match
-          }
-
-          nickname = nickname ?? ''
-          const possibleMentions = channelParticipants
-            .filter(
-              user =>
-                user.nickname.startsWith(nickname) && !channelParticipants.find(user => user.nickname === nickname)
-            )
-            .filter(isDefined)
-
-          if (JSON.stringify(mentionsToSelect) !== JSON.stringify(possibleMentions)) {
-            setMentionsToSelect(possibleMentions)
-            setTimeout(() => {
-              setSelected(0)
-            }, 0)
-          }
-
-          // Wrap mention in spans to be able to treat it as an anchor for popper
-          console.log('Returning wrapped mentions', nickname)
-          return `<span>@${nickname}</span>`
-        }
-      )
-
-      return result
-    },
-    [mentionsToSelect, setMentionsToSelect]
-  )
-
-  const sanitizedHtml = findMentions(htmlMessage)
+  const adjustTextAreaHeight = (el: HTMLTextAreaElement) => {
+    // Workaround for making textarea's height adapt to the content
+    el.style.height = ''
+    if (el.scrollHeight > el.clientHeight) {
+      el.style.height = `${el.scrollHeight}px`
+    }
+  }
 
   const caretLineTraversal = (focusLine: Node | null | undefined, anchorLinePosition = 0) => {
     if (!focusLine?.nodeValue) return
@@ -364,20 +287,13 @@ export const ChannelInputComponent: React.FC<ChannelInputProps> = ({
   }
 
   const onChangeCb = useCallback(
-    (e: ContentEditableEvent) => {
+    (e: React.ChangeEvent<HTMLTextAreaElement>) => {
       if (inputState === INPUT_STATE.AVAILABLE) {
-        // @ts-expect-error
-        setMessage(e.nativeEvent.target.innerText)
-        // @ts-expect-error
-        if (!e.nativeEvent.target.innerText) {
-          setHtmlMessage('')
-        } else {
-          setHtmlMessage(e.target.value)
-        }
+        setMessage(e.target.value)
+        adjustTextAreaHeight(e.target)
       }
-      setAnchorEl(e.currentTarget.lastElementChild)
     },
-    [setAnchorEl, onChange, setHtmlMessage]
+    [onChange]
   )
 
   const inputStateRef = React.useRef(inputState)
@@ -385,51 +301,8 @@ export const ChannelInputComponent: React.FC<ChannelInputProps> = ({
     inputStateRef.current = inputState
   })
 
-  const mentionSelectAction = (e: React.MouseEvent | React.KeyboardEvent) => {
-    e.preventDefault()
-    if (!mentionsToSelectRef.current || !refSelected.current) return
-    const nickname = mentionsToSelectRef.current[refSelected.current].nickname
-    setHtmlMessage(htmlMessage => {
-      const wrapped = `<span class="${classes.highlight}">@${nickname}</span>&nbsp;`
-      return htmlMessage.replace(/<span>[^/]*<\/span>$/g, wrapped)
-    })
-    // Replace mentions characters with full nickname in original message string
-    setMessage(message.replace(/(\b(\w+)$)/, `${nickname} `))
-    // Clear popper items after choosing mention
-    setMentionsToSelect([])
-    inputRef.current?.el.current.focus()
-  }
-
   const onKeyDownCb = useCallback(
     (e: React.KeyboardEvent) => {
-      if (!isRefSelected(refSelected.current)) {
-        throw new Error('refSelected is on unexpected type')
-      }
-      if (!mentionsToSelectRef?.current) {
-        return
-      }
-      if (mentionsToSelectRef.current.length) {
-        if (e.nativeEvent.keyCode === 40) {
-          if (refSelected.current + 1 >= mentionsToSelectRef.current.length) {
-            setSelected(0)
-          } else {
-            setSelected(refSelected.current + 1)
-          }
-          e.preventDefault()
-        }
-        if (e.nativeEvent.keyCode === 38) {
-          if (refSelected.current - 1 < 0) {
-            setSelected(mentionsToSelectRef.current.length - 1)
-          } else {
-            setSelected(refSelected.current - 1)
-          }
-          e.preventDefault()
-        }
-        if (e.nativeEvent.keyCode === 13 || e.nativeEvent.keyCode === 9) {
-          mentionSelectAction(e)
-        }
-      }
-
       if (e.key === 'ArrowDown') {
         const anchorNode: Node | null | undefined = window?.getSelection()?.anchorNode
 
@@ -481,16 +354,16 @@ export const ChannelInputComponent: React.FC<ChannelInputProps> = ({
         caretLineTraversal(previousNode, window?.getSelection()?.anchorOffset)
       }
 
-      if (e.nativeEvent.keyCode === 13) {
+      if (e.nativeEvent.key === 'Enter') {
         if (e.shiftKey) {
           // Accept this input for additional lines in the message box
         } else if (inputStateRef.current === INPUT_STATE.AVAILABLE) {
           e.preventDefault()
           const target = e.target as HTMLInputElement
-          onChange(target.innerText)
-          onKeyPress(target.innerText)
+          onChange(target.value)
+          onKeyPress(target.value)
           setMessage('')
-          setHtmlMessage('')
+          target.style.height = ''
         } else {
           e.preventDefault()
           if (infoClass !== classNames(classes.backdrop, classes.blinkAnimation)) {
@@ -500,18 +373,7 @@ export const ChannelInputComponent: React.FC<ChannelInputProps> = ({
         }
       }
     },
-    [
-      inputState,
-      message,
-      mentionsToSelectRef,
-      onChange,
-      onKeyPress,
-      setMessage,
-      setHtmlMessage,
-      infoClass,
-      setInfoClass,
-      setSelected,
-    ]
+    [inputState, message, onChange, onKeyPress, setMessage, infoClass, setInfoClass]
   )
 
   const handleFileInput = (event: React.ChangeEvent<HTMLInputElement>) => {
@@ -535,22 +397,6 @@ export const ChannelInputComponent: React.FC<ChannelInputProps> = ({
         direction='column'
         justifyContent='center'
       >
-        {/* <MentionPoper anchorEl={anchorEl} selected={selected}>
-          {mentionsToSelect.map((target, index) => (
-            <MentionElement
-              key={index}
-              name={target.nickname}
-              highlight={index === selected}
-              onMouseEnter={() => {
-                setSelected(index)
-              }}
-              channelName={channelName}
-              onClick={e => {
-                mentionSelectAction(e)
-              }}
-            />
-          ))}
-        </MentionPoper> */}
         <Grid
           container
           direction='row'
@@ -577,37 +423,31 @@ export const ChannelInputComponent: React.FC<ChannelInputProps> = ({
               justifyContent='center'
               alignItems='center'
             >
-              <Grid item xs className={classes.textArea}>
-                <ContentEditable
-                  ref={inputRef}
-                  placeholder={`Message ${inputPlaceholder}`}
-                  className={classes.input}
-                  onClick={() => {
-                    if (!focused) {
-                      setFocused(true)
-                    }
-                  }}
-                  disabled={inputState !== INPUT_STATE.AVAILABLE}
-                  html={sanitizedHtml}
-                  onChange={onChangeCb}
-                  onKeyDown={onKeyDownCb}
-                  onPaste={async e => {
-                    e.preventDefault()
-                    const files = e.clipboardData.files
-                    for (let i = 0; i < files.length; i++) {
-                      const fileExt = path.extname(files[i].name).toLowerCase()
-                      const fileName = path.basename(files[i].name, fileExt)
-                      const arrayBuffer = await files[i].arrayBuffer()
-                      handleClipboardFiles(arrayBuffer, fileExt, fileName)
-                    }
-                    if (!files.length) {
-                      const text = e.clipboardData.getData('text/plain').replace(/(\r\n|\r|\n)/g, '<br />')
-                      document.execCommand('insertHTML', false, text)
-                    }
-                  }}
-                  data-testid='messageInput'
-                />
-              </Grid>
+              <textarea
+                ref={textAreaRef}
+                placeholder={`Message ${inputPlaceholder}`}
+                className={classes.input}
+                onClick={() => {
+                  if (!focused) {
+                    setFocused(true)
+                  }
+                }}
+                value={message}
+                disabled={inputState !== INPUT_STATE.AVAILABLE}
+                onChange={onChangeCb}
+                onKeyDown={onKeyDownCb}
+                onPaste={async e => {
+                  const files = e.clipboardData.files
+                  if (files.length) e.preventDefault()
+                  for (let i = 0; i < files.length; i++) {
+                    const fileExt = path.extname(files[i].name).toLowerCase()
+                    const fileName = path.basename(files[i].name, fileExt)
+                    const arrayBuffer = await files[i].arrayBuffer()
+                    handleClipboardFiles(arrayBuffer, fileExt, fileName)
+                  }
+                }}
+                data-testid='messageInput'
+              />
               {children}
               <div className={classes.icons}>
                 <Grid item className={classes.actions}>
@@ -662,7 +502,6 @@ export const ChannelInputComponent: React.FC<ChannelInputProps> = ({
                       <div data-testid={'emoji-picker'} className={classes.picker}>
                         <Picker
                           onEmojiClick={(emojiData, _event) => {
-                            setHtmlMessage(htmlMessage => htmlMessage + emojiData.emoji)
                             setMessage(message + emojiData.emoji)
                             setOpenEmoji(false)
                           }}