+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-[unreleased]
+[2.0.3-alpha.0]
+
+* Filter CSRs - remove old csrs and replace with new for each pubkey
 
 * Fixed mobile bugs - joining by QR code and not showing username taken screen for user who has unique name
 
@@ -11,6 +13,10 @@
 * Removed registration attempts selector and corresponding usage.
 
 * Revert adjusting bootstrap scripts for developing on Windows
+
+* Channel input - replaced ContentEditable with textarea
+
+* Fix - up/down arrows now work properly inside channel input (textarea)
 
 [2.0.1-alpha.2]